Address crud operation included

+++ src/main/java/com/don8/repository/UserRepository.java
@@ -1,13 +1,12 @@
 package com.don8.repository;
 
-import com.don8.model.User;
+import com.don8.model.dbentity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
